refactor: parentButton 컴포넌트 따로 분리 (#41)
